django_backend/backend/authentication/serializers.py
from rest_framework import serializers
from .models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['id', 'email', 'full_name', 'phone_number', 'role', 'created_at']
        read_only_fields = ['id', 'created_at']


class RegisterSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, min_length=8)

    class Meta:
        model = CustomUser
        fields = ['email', 'password', 'full_name', 'phone_number']

    def create(self, validated_data):
        validated_data['role'] = CustomUser.Role.RECRUITER
        user = CustomUser.objects.create_user(**validated_data)
        return user
